requirements.txt
# Bot packages
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
debugpy==1.6.7
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
Pillow==9.5.0
python-dotenv==1.0.0
python-telegram-bot==20.0
requests==2.29.0
rfc3986==1.5.0
sniffio==1.3.0
urllib3==1.26.15

# User Packages
paho-mqtt==1.6.1